I added different types of noise.
This works without errors for me. I placed it so that the half without the gridlines would first get saved as normal, and then multiple copies of it with different noise levels will be saved. We should probably discuss the logistics of the placement of the special effects later. I can also easily blur some images if need be.